+++ tools/ui/src/lib/components/app/models/discover/ModelsDiscoverList.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import ModelsDiscoverItem from './ModelsDiscoverItem.svelte';
+	import ModelsDiscoverListItem from './ModelsDiscoverListItem.svelte';
 	import type { HfModelInfo } from '$lib/types/huggingface';
 
 	interface Props {
@@ -15,6 +15,11 @@
 
 <ul class="space-y-0.5 p-2">
 	{#each models as model (model.id)}
-		<ModelsDiscoverItem active={model.id === activeId} {model} {onSelect} {showBaseModelAvatar} />
+		<ModelsDiscoverListItem
+			active={model.id === activeId}
+			{model}
+			{onSelect}
+			{showBaseModelAvatar}
+		/>
 	{/each}
 </ul>
