+++ tools/ui/src/lib/components/app/models/discover/ModelsDiscoverListItem.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import ModelId from '../ModelId.svelte';
+	import ModelsDiscoverAvatar from './ModelsDiscoverAvatar.svelte';
 	import { isAuxSidecar, type ModelSidecar } from '$lib/constants';
 	import { HuggingFaceService, ModelsService } from '$lib/services';
 	import { modelsHubStore } from '$lib/stores';
@@ -10,9 +11,25 @@
 
 	interface Props {
 		model: HfModelInfo;
+		active?: boolean;
+		/** Show the original (base) model's org avatar instead of the repo's org. */
+		showBaseModelAvatar?: boolean;
+		onSelect?: (modelId: string) => void;
 	}
 
-	let { model }: Props = $props();
+	let { active = false, model, onSelect, showBaseModelAvatar = false }: Props = $props();
+
+	let org = $derived(model.id.split('/')[0] ?? model.id);
+
+	// Org whose avatar is shown: the base model's org when showBaseModelAvatar
+	// (e.g. the Qwen logo for ggml-org/Qwen3.8-27B-GGUF), else the repo's org.
+	let avatarOrg = $derived.by(() => {
+		if (!showBaseModelAvatar) return org;
+
+		const base = HuggingFaceService.getBaseModels(model)[0];
+
+		return base?.split('/')[0] || org;
+	});
 
 	let contextLength = $derived(model.gguf?.context_length);
 
@@ -130,19 +147,32 @@
 	});
 </script>
 
-<span class="min-w-0 flex-1">
-	<ModelId
-		class="min-w-0"
-		{contextLength}
-		{draftSidecars}
-		hideOrgName
-		iconsOnNewLine
-		{modalities}
-		modelId={model.id}
-		params={paramsFallback}
-		{sizeRange}
-		{supportsThinking}
-		{supportsToolUse}
-		wrap
-	/>
-</span>
+<li>
+	<button
+		aria-current={active ? 'page' : undefined}
+		class="flex w-full cursor-pointer items-start gap-2.5 rounded-lg p-2.5 text-left transition-colors {active
+			? 'bg-primary/10 hover:bg-primary/15'
+			: 'hover:bg-muted/60'}"
+		onclick={() => onSelect?.(model.id)}
+		type="button"
+	>
+		<ModelsDiscoverAvatar org={avatarOrg} quantOrg={showBaseModelAvatar ? org : undefined} />
+
+		<span class="min-w-0 flex-1">
+			<ModelId
+				class="min-w-0"
+				{contextLength}
+				{draftSidecars}
+				hideOrgName
+				iconsOnNewLine
+				{modalities}
+				modelId={model.id}
+				params={paramsFallback}
+				{sizeRange}
+				{supportsThinking}
+				{supportsToolUse}
+				wrap
+			/>
+		</span>
+	</button>
+</li>
